First stage or porting the rotational delta arm for FirePick to Smoothie

+++ src/main.cpp
@@ -110,7 +110,7 @@ void init() {
     kernel->streams->printf("  Build version %s, Build date %s\r\n", version.get_build(), version.get_build_date());
 
     //some boards don't have leds.. TOO BAD!
-    kernel->use_leds= !kernel->config->value( disable_leds_checksum )->by_default(false)->as_bool();
+    kernel->use_leds= !kernel->config->value( (unsigned int)disable_leds_checksum )->by_default(false)->as_bool();
 
     bool sdok= (sd.disk_initialize() == 0);
     if(!sdok) kernel->streams->printf("SDCard is disabled\r\n");
@@ -201,11 +201,11 @@ void init() {
 #endif
 
     kernel->add_module( &usbserial );
-    if( kernel->config->value( second_usb_serial_enable_checksum )->by_default(false)->as_bool() ){
+    if( kernel->config->value( (unsigned int)second_usb_serial_enable_checksum )->by_default(false)->as_bool() ){
         kernel->add_module( new(AHB0) USBSerial(&u) );
     }
 
-    if( kernel->config->value( dfu_enable_checksum )->by_default(false)->as_bool() ){
+    if( kernel->config->value( (unsigned int)dfu_enable_checksum )->by_default(false)->as_bool() ){
         kernel->add_module( new(AHB0) DFU(&u));
     }
     kernel->add_module( &u );
